Pbr render demo

+++ demo/demos/08_DynamicBufferDemo.cpp
@@ -73,7 +73,7 @@ namespace mygfx::demo {
 
 				float2 center = { linearRand<float>(0, width), linearRand<float>(0.0f, height) };
 				float halfSize = linearRand<float>(5.0f, 10.0f);
-				int index = mTextures[linearRand<int>(0, mTextures.size() - 1)]->index();
+				int index = mTextures[linearRand<int>(0, (int)mTextures.size() - 1)]->index();
 				mSprites.push_back(
 					Sprite2D{ center, {halfSize, halfSize},
 					0.0f, 0xffffffff, Rect{ {0, 0}, {1, 1} }, index, {linearRand<float>(0, width), linearRand<float>(0.0f, height)} });
